requirements.txt
pyspark==3.5.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.13.0
jupyter>=1.0.0
kaggle>=1.6.0
pyarrow>=14.0.0
folium>=0.14.0
notebook>=7.0.0
kafka-python>=2.0.0